refactor: moving `HintBar` into widgets module
