+++ src/components/mod.rs
@@ -1,15 +1,13 @@
-use tui::{backend::Backend, Frame};
-
-use hint_bar::HintBar;
-
 pub mod app;
-pub mod hint_bar;
 pub mod metadata;
 pub mod notification;
 pub mod skimmer;
 pub mod todo_input;
 pub mod todo_list;
 mod utils;
+
+use crate::widgets::hint_bar::HintBar;
+use tui::{backend::Backend, Frame};
 
 pub trait Component {
     fn draw<B: Backend>(&mut self, _f: &mut Frame<B>) {}
