+++ src/components/utils.rs
@@ -43,10 +43,3 @@ pub fn default_block(title: &'static str) -> Block {
         .borders(Borders::ALL)
         .title(title)
 }
-
-fn match_char(data: &char) -> bool {
-    match *data {
-        '\x01'..='\x08' | '\u{10FFFE}'..='\u{10FFFF}' => true,
-        _ => false,
-    }
-}
